Rename views.py to reportviews.py
