+++ networkx/classes/__init__.py
@@ -2,8 +2,10 @@ from .graph import Graph
 from .digraph import DiGraph
 from .multigraph import MultiGraph
 from .multidigraph import MultiDiGraph
-from .views import *
+from .reportviews import *
 from .graphviews import *
 from .ordered import *
 
 from .function import *
+
+import coreviews
