+++ networkx/classes/graph.py
@@ -23,7 +23,7 @@ from collections import Mapping
 
 import networkx as nx
 from networkx.classes.coreviews import AtlasView, AdjacencyView
-from networkx.classes.views import NodeView, EdgeView, DegreeView
+from networkx.classes.reportviews import NodeView, EdgeView, DegreeView
 from networkx.exception import NetworkXError
 import networkx.convert as convert
 from networkx.utils import pairwise
